个人知识库问答系统/requirements.txt
# 核心依赖
langchain==0.1.0
langchain-community==0.0.10

# 向量数据库
faiss-cpu==1.12.0

# 文档处理
pypdf==4.2.0
python-docx==1.1.0

# 文本嵌入
sentence-transformers==2.2.2

# 可选：本地模型支持
ollama==0.1.9

# 基础工具
numpy==1.26.4

sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0